backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.11.2
black==24.4.2
blessed==1.20.0
boto3==1.26.38
botocore==1.29.38
certifi==2022.12.7
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
constantly==23.10.4
crispy-bootstrap4==2022.1
cryptography==43.0.3
daphne==4.1.2
dataclasses-json==0.6.7
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.16
django-bootstrap-datepicker-plus==5.0.3
django-bootstrap4==22.3
django-cleanup==8.0.0
django-cors-headers==4.6.0
django-crispy-forms==2.0
django-mathfilters==1.0.0
django-multiselectfield==0.1.12
django-storages==1.13.2
django-tinymce==3.6.1
django-widget-tweaks==1.4.12
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
docutils==0.17.1
editor==1.6.6
email_validator==2.2.0
faiss-cpu==1.8.0.post1
frozenlist==1.4.1
fsrs==3.1.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
incremental==24.7.2
inquirer==3.4.0
Jinja2==3.0.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.24
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.93
Levenshtein==0.25.1
MarkupSafe==2.1.5
marshmallow==3.22.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numerize==0.12
numpy==1.26.4
openai==1.37.1
orjson==3.10.7
outcome==1.3.0.post0
packaging==24.0
pathspec==0.12.1
Pillow==9.3.0
platformdirs==4.2.2
psycopg2==2.9.10
psycopg2-binary==2.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.10.5
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0.2
rapidfuzz==3.9.7
readchar==4.2.0
redis==5.0.1
regex==2024.7.24
reportlab==4.2.2
requests==2.28.2
runs==1.2.2
s3transfer==0.6.0
selenium==4.9.1
service-identity==24.2.0
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4
Sphinx==1.8.5
sphinx-rtd-theme==0.4.3
sphinxcontrib-newsfeed==0.1.4
sphinxcontrib-serializinghtml==2.0.0
sphinxcontrib-websupport==1.2.4
SQLAlchemy==2.0.35
sqlparse==0.4.3
stripe==5.2.0
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
Twisted==24.10.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==1.26.13
wcwidth==0.2.13
webdriver-manager==4.0.2
websockets==11.0.3
whitenoise==6.2.0
wsproto==1.2.0
xmod==1.8.1
yarl==1.11.1
zope.interface==7.1.1
